Дописал функцию вызова данных с сервера
