+++ js/api.js
@@ -1,29 +1,33 @@
+import { showAlert } from './utils.js';
+
 const URL_GET_DATA = 'https://23.javascript.pages.academy/kekstagram/data';
 const URL_SEND_DATA = 'https://23.javascript.pages.academy/kekstagram';
 
-const getData = (onFail) => fetch(URL_GET_DATA)
-  .then((response) => {
-    if (response.ok) {
-      return response.json();
-    }
-    throw new Error(`${response.status} — ${response.statusText}`);
-  })
-  .catch(onFail);
+const onResponseCallback = (response) => {
+  if (response.ok) {
+    return response.json();
+  }
+  throw new Error(`${response.status} — ${response.statusText}`);
+};
 
-const sendData = (onSuccess, onFail, body) => {
+const getData = (onSuccess) => fetch(URL_GET_DATA)
+  .then(onResponseCallback)
+  .then(onSuccess())
+  .catch(() => {
+    showAlert('Не удалось получить данные с сервера. Попробуйте ещё раз.');
+  });
+
+const sendData = (onSuccess, body) => {
   fetch(URL_SEND_DATA,
     {
       method: 'POST',
       body,
     },
   )
-    .then((response) => {
-      if (response.ok) {
-        return onSuccess();
-      }
-      throw new Error(`${response.status} — ${response.statusText}`);
-    })
-    .catch(onFail);
+    .then(onResponseCallback)
+    .then(onSuccess())
+    .catch(() => {
+      showAlert();
+    });
 };
-
 export { getData, sendData };
